Add new utils for handling partnerships and rewards
Introduces several new utility files that handle partnerships and rewards for individual campaigns. Each file contains multiple hooks performing distinct functionalities.

- `use-start-partnership.tsx`: Contains the logic to start a new partnership. It handles any exceptions that may occur and notifies the user regarding the partnership's success or failure.

- `claims-and-rewards.tsx`: Contains hooks to handle rewards claiming by users. It makes sure users are aware of the rewards they receive, keeps track of rewards claimed, and handles any possible exceptions here too.

- `funds.ts`: Handles the funds staked into a campaign. It monitors the campaign longevity and recalculates funds.

- `partnership-hooks.tsx`: Contains hooks to manage partnerships for the campaign. It handles querying for partnership metrics and filters partnerships that match the specific campaign.

The changes also include restructuring engagement with the components in `campaign/[chain]/[id].tsx`. The new component structure makes use of the aforementioned hooks for cleaner, easier-to-follow code.

These changes are implemented to increase modularity, making it easier to track, manage, and debug different parts of the campaign mechanism.
